+++ src/core/xml_parser.js
